ENH: Adding kinematicParcelFoam and tutorial
The original thermoSurfaceFilm sub-models was devided between kinematicSurfaceFilm
and thermoSurfaceFilm in order to use the surfaceFilm model in a
kinematicCloud.

The film interaction models are now in a kinematicSurface class which can
be used in a kinematic cloud adding constant thermal properties(p and T)
for some sub-models. i.e. drySplashInteraction/wetSplashInteraction

pRef and Tref were added to the kinematicSurfaceFilm as entry to the
regionFilm when used with a kinematic cloud.

In the finite area surface film model Tref, pRef are stored in filmSubModelBase.